extract only props
